nuri-gpt-backend/requirements.txt
# FastAPI 기본 의존성
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.10.0
email-validator>=2.0.0  # pydantic EmailStr 사용
pydantic-settings>=2.1.0

# 환경 변수 관리
python-dotenv>=1.0.0

# Supabase 연동
supabase~=2.4.6

# 비동기 HTTP 클라이언트
httpx>=0.27.0

# 테스트
pytest>=7.4.3
pytest-asyncio>=0.21.1

# CORS 미들웨어
python-multipart>=0.0.6

# AI 및 문서 처리
google-generativeai>=0.8.0
Pillow>=10.0.0  # 이미지 처리용
requests
